requirements.txt
mcp
azure-devops
pydantic
python-dotenv
